requirements.txt
asgiref==3.8.1
black==24.10.0
bleach==6.1.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==5.1.1
django-browser-reload==1.15.0
django-cors-headers==4.4.0
django-tailwind==3.8.0
djangorestframework==3.15.2
filelock==3.16.1
identify==2.6.1
idna==3.10
Markdown==3.5.2
martor==1.6.44
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.6
webencodings==0.5.1
